Create a BaseUploadServe inside the core The AlfrescoUpload should extend the BaseUpload
