+++ ng2-components/ng2-alfresco-core/src/models/file.model.ts
