+++ ng2-components/ng2-alfresco-core/src/models/index.ts
@@ -16,3 +16,4 @@
  */
 
 export * from './card-view.model';
+export * from './file.model';
