+++ ng2-components/ng2-alfresco-core/src/services/index.ts
@@ -32,3 +32,4 @@ export * from './alfresco-translation.service';
 export * from './alfresco-translate-loader.service';
 export * from './app-config.service';
 export * from './thumbnail.service';
+export * from './base-upload.service';
