+++ ng2-components/ng2-alfresco-upload/README.md
@@ -66,7 +66,7 @@ npm install ng2-alfresco-upload
 
 ```html
 <adf-upload-button 
-    [rootFolderId]="-my-"
+    [parentId]="-my-"
     [uploadFolders]="true"
     [multipleFiles]="false"
     [acceptedFilesType]=".jpg,.gif,.png,.svg"
@@ -86,7 +86,9 @@ npm install ng2-alfresco-upload
 | multipleFiles | boolean | false | Allow/disallow multiple files |
 | acceptedFilesType | string | * |  array of allowed file extensions , example: ".jpg,.gif,.png,.svg" |
 | **(deprecated)** currentFolderPath | string | '/Sites/swsdp/documentLibrary' | define the path where the files are uploaded. **Deprecated in 1.6.0: use rootFolderId instead.** |
-| rootFolderId | string | '-root-' | The ID of the root folder node. |
+| **(deprecated)** rootFolderId | string | '-root-' | The ID of the root folder node.
+**Deprecated in 1.6.2: use parentId instead.** |
+| parentId | string | empty | The ID of the root. It can be the nodeId if you are using the upload for the Content Service. |
 | versioning | boolean | false | Versioning false is the default uploader behaviour and it rename using an integer suffix if there is a name clash. Versioning true to indicate that a major version should be created |
 | staticTitle | string | (predefined) | define the text of the upload button |
 | disableWithNoPermission | boolean | false |  If the value is true and the user doesn't have the permission to delete the node the button will be disabled |
