+++ ng2-components/ng2-alfresco-upload/index.ts
@@ -32,9 +32,7 @@ export * from './src/components/upload-drag-area.component';
 export * from './src/services/upload.service';
 export * from './src/directives/file-draggable.directive';
 export * from './src/components/file-uploading-list.component';
-export * from './src/models/file.model';
 export * from './src/models/permissions.model';
-export * from './src/events/file.event';
 
 export const UPLOAD_DIRECTIVES: any[] = [
     FileDraggableDirective,
