+++ ng2-components/ng2-alfresco-upload/src/components/file-uploading-dialog.component.ts
@@ -16,9 +16,7 @@
  */
 
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
-import { AlfrescoTranslationService } from 'ng2-alfresco-core';
-import { FileUploadCompleteEvent } from '../events/file.event';
-import { FileModel, FileUploadStatus } from '../models/file.model';
+import { AlfrescoTranslationService, FileUploadCompleteEvent, FileUploadStatus } from 'ng2-alfresco-core';
 import { UploadService } from '../services/upload.service';
 
 @Component({
