+++ ng2-components/ng2-alfresco-upload/src/components/upload-button.component.ts
@@ -17,9 +17,8 @@
 
 import { Component, ElementRef, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
 import { MinimalNodeEntryEntity } from 'alfresco-js-api';
-import { AlfrescoApiService, AlfrescoContentService, AlfrescoTranslationService, FileUtils, LogService, NotificationService } from 'ng2-alfresco-core';
+import { AlfrescoApiService, AlfrescoContentService, AlfrescoTranslationService, FileModel, FileUtils, LogService, NotificationService } from 'ng2-alfresco-core';
 import { Observable, Subject } from 'rxjs/Rx';
-import { FileModel } from '../models/file.model';
 import { PermissionModel } from '../models/permissions.model';
 import { UploadService } from '../services/upload.service';
